probando grups id

+++ app/Http/Controllers/TiendaController.php
@@ -5,10 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Groups;
 use App\Models\OrdenPurchases;
 use App\Models\Packages;
+use App\Models\User;
 use Illuminate\Support\Facades\View;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\DB;
+
 
 
 class TiendaController extends Controller
@@ -19,8 +22,6 @@ class TiendaController extends Controller
     public function __construct()
     {
         $this->apis_key_nowpayments = 'YH0WTN1-5T64QQC-MRVZZPE-0DSX41R';
-         //la tienda funciona pero la api key de nowpaymenst no esta activa
-        // con mi api key si funciona YH0WTN1-5T64QQC-MRVZZPE-0DSX41R
     }
 
     /**
@@ -55,11 +56,11 @@ class TiendaController extends Controller
             View::share('titleg', 'Tienda - Productos');
             $category = Groups::find($idgroup);
             $services = $category->getPackage->where('status', 1);
-dd($idgroup);
+
             return view('shop.products', compact('services'));
         } catch (\Throwable $th) {
             Log::error('Tienda - products -> Error: '.$th);
-            abort(403, "Ocurrio un error en products funcion , contacte con el administrador");
+            abort(403, "Ocurrio un error en shop, contacte con el administrador");
         }
     }
 
@@ -71,47 +72,50 @@ dd($idgroup);
      */
     public function procesarOrden(Request $request)
     {
-        $validate = $request->validate([
-            'idproduct' => 'required'
-        ]);
 
-        try {
-            if ($validate) {
-                $paquete = Packages::find($request->idproduct);
+        //Obtener el valor de la direccion de billetera del usuario logueado
+        $check_wallet = DB::table('users')
+            ->where('id', '=', Auth::user()->id)->value('wallet_address');
 
-                $porcentaje = ($paquete->price * 0.03);
-                $total = ($paquete->price + $porcentaje);
+        if($check_wallet != null || $check_wallet != ''){//Validar si usuario tiene la wallet registrada.
 
-                $data = [
-                    'iduser' => Auth::id(),
-                    'group_id' => $paquete->getGroup->id,
-                    'package_id' => $paquete->id,
-                    'cantidad' => 1,
-                    'total' => $total
-                ];
-                dd($data);
+            $validate = $request->validate([
+                'idproduct' => 'required'
+            ]);
 
-                $data['idorden'] = $this->saveOrden($data);
+            try {
+                if ($validate) {
+                    $paquete = Packages::find($request->idproduct);
 
-                $data['descripcion'] = $paquete->description;
+                    $porcentaje = ($paquete->price * 0.03);
+                    $total = ($paquete->price + $porcentaje);
 
-                $url = $this->generalUrlOrden($data);
+                    $data = [
+                        'iduser' => Auth::id(),
+                        'group_id' => $paquete->getGroup->id,
+                        'package_id' => $paquete->id,
+                        'cantidad' => 1,
+                        'total' => $total
+                    ];
+                    $data['idorden'] = $this->saveOrden($data);
 
-                if (!empty($url)) {
-                    return redirect($url);
+                    $data['descripcion'] = $paquete->description;
+                    $url = $this->generalUrlOrden($data);
+                    if (!empty($url)) {
+                        return redirect($url);
 
-                }else{
-
-                   OrdenPurchases::where('id', $data['idorden'])->delete();
-                   return redirect()->back()->with('msj-info', 'Problemas al general la orden, intente mas tarde');
+                    }else{
+                        OrdenPurchases::where('id', $data['idorden'])->delete();
+                        return redirect()->back()->with('msj-info', 'Problemas al general la orden, intente mas tarde');
+                    }
                 }
-
-
+            } catch (\Throwable $th) {
+               Log::error('Tienda - procesarOrden -> Error: '.$th);
+                abort(403, "Ocurrio un error en procesa rorden function , contacte con el administrador");
             }
-        } catch (\Throwable $th) {
-            Log::error('Tienda - procesarOrden -> Error: '.$th);
-            abort(403, "Ocurrio un error (1) , contacte con el administrador");
         }
+        return redirect()->back()->with('msj-warning', 'Necesita registrar su billetera para poder continuar con la operación.');
+
     }
 
     /**
@@ -122,7 +126,10 @@ dd($idgroup);
      */
     public function saveOrden($data): int
     {
+       // dd($data);
+
         return OrdenPurchases::insertGetId($data);
+
     }
 
     /**
@@ -163,6 +170,7 @@ dd($idgroup);
                 'x-api-key: '.$this->apis_key_nowpayments,
                 'Content-Type:application/json'
             ];
+           // dd($headers);
             $resul = '';
             $curl = curl_init();
 
@@ -192,8 +200,6 @@ dd($idgroup);
 
                 $response = curl_exec($curl);
                 $err = curl_error($curl);
-              //  dd($dataRaw);
-
                 curl_close($curl);
                 if ($err) {
                     Log::error('Tienda - generalUrlOrden -> Error curl: '.$err);
@@ -206,7 +212,7 @@ dd($idgroup);
             return $resul;
         } catch (\Throwable $th) {
             Log::error('Tienda - generalUrlOrden -> Error: '.$th);
-            abort(403, "Ocurrio un error, contacte con el administrador");
+            abort(403, "Ocurrio un error2, contacte con el administrador");
         }
     }
 }
